requirements.txt
snowflake-connector-python
pandas